requirements.txt
appdirs>=1.4.3
asgiref>=3.5.2
backports.zoneinfo>=0.2.1
beautifulsoup4>=4.11.1
CacheControl>=0.12.6
certifi>=2019.11.28
chardet>=3.0.4
colorama>=0.4.3
contextlib2>=0.6.0
distlib>=0.3.0
distro>=1.4.0
Django>=4.0.5
django-appconf>=1.0.4
django-bootstrap-v5>=1.0.11
django-ckeditor>=6.0.0
django-cleanup>=5.1.0
django-clearcache>=1.2.1
django-debug-toolbar>=3.2.3
django-imagekit>=4.1.0
django-js-asset>=1.2.2
html5lib>=1.0.1
idna>=2.8
ipaddr>=2.2.0
lockfile>=0.12.2
msgpack>=0.6.2
mysqlclient>=2.0.1
packaging>=20.3
pep517>=0.8.2
pilkit>=2.0
Pillow>=8.0.1
progress>=1.5
psycopg2>=2.8.6
Pygments>=2.12.0
pyparsing>=2.4.6
python-monkey-business>=1.0.0
pytoml>=0.1.21
pytz>=2020.4
requests>=2.22.0
retrying>=1.3.3
six>=1.14.0
soupsieve>=2.3.2.post1
sqlparse>=0.4.1
urllib3>=1.25.8
webencodings>=0.5.1
